[change]timelineを分けて書き直した
