+++ app/domain/repository/timeline.go
@@ -1,0 +1,10 @@
+package repository
+
+import (
+	"context"
+	"yatter-backend-go/app/domain/object"
+)
+
+type Timeline interface {
+	GetPublicTimeline(ctx context.Context, maxID, sinceID, limit int) ([]*object.Status, error)
+}
